issue #101 - refactored StyleSheet.create(). Moved OS validation logic into StyleSheet.create
